requirements.txt
absl-py==2.2.2
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
audioread==3.0.1
beautifulsoup4==4.13.4
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==3.1.1
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
demucs==4.0.1
dora_search==0.1.12
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.18.0
flatbuffers==1.12
free_proxy==1.1.3
freeproxy==1.0.2
fsspec==2025.3.2
gast==0.4.0
gevent==24.11.1
google-auth==2.39.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
julius==0.2.7
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.9.0
Keras-Preprocessing==1.1.2
lameenc==1.8.1
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.39.1
lxml==5.3.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.1.0
mutagen==1.47.0
namex==0.0.9
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
openpyxl==3.1.5
openunmix==1.3.0
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.17.9
pexpect==4.9.0
platformdirs==4.3.7
pooch==1.8.2
prompt_toolkit==3.0.51
protobuf==3.19.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.22.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytube==15.0.0
pytubefix==8.12.3
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rich==14.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
stack-data==0.6.3
submitit==1.5.2
sympy==1.14.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.16.2
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.9.0
tensorflow-metal==1.2.0
termcolor==3.1.0
threadpoolctl==3.6.0
torch==2.7.0
torchaudio==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
treetable==0.2.5
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yt-dlp==2025.3.31
ytmusicapi==1.10.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
